requirements.txt
kucoin-python==1.0.8
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
ta==0.10.2
pytest==7.4.2
black==23.7.0
flake8==6.1.0
joblib==1.3.2
pybit==5.5.0 